WIP build of specific user page
