+++ package.json
@@ -6,6 +6,7 @@
     "express": "^4.15.4",
     "react": "^15.6.1",
     "react-dom": "^15.6.1",
+    "react-router-dom": "^4.2.2",
     "react-scripts": "1.0.13",
     "web3": "^1.0.0-beta.20"
   },
